core/backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.10
python-dotenv==1.0.0
bcrypt==4.0.1
Werkzeug==2.3.7
gunicorn==21.2.0
email-validator==2.0.0
redis==5.0.1
python-socketio==5.11.0
APScheduler==3.10.4
Pillow==10.1.0
